+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased] - yyyy-mm-dd
 ### Added
-- New primary-navigation component
+- Header has the ability to contain a navigation bar
 
 ## [1.2.1] - 2023-02-21
 ### Changed